replace middleware with Interceptor to solve the always anonymous user problem: (the middleware was running before the guards) add exclude field to the abstract repo findone function (solve the problem of returning the user email in the get('userId') endpoint in the user-service) fix issue with update me in user-service add vscode tasks : for local development i dont like to make a docker image for my services so i run them locally in the terminal so instead of looping on each service and run npm start just ctrl+shit+p ->  Tasks: Run Task -> dev:all-services fix: mimetype always null , cause(cloudinary response has no attr like that instead i got it from the original file)

+++ apps/users-service/src/services/users-service.service.ts
@@ -23,6 +23,12 @@ export class UsersService
 
   async createUser(userDto:UserProfileDto, photoMetadata: PhotoMetadataDto) 
   {
+    this.logger.logInfo({
+      functionName: 'createUser',
+      message: `KAFKA event received for userId: ${userDto.userId}, email: ${userDto.email}`,
+      userId: userDto.userId,
+      additionalData: { email: userDto.email }
+    });
     const user = new User({
       ...userDto,
     });
@@ -45,9 +51,9 @@ export class UsersService
     return this.UserRepo.create(user);
   }
 
-  async getUserById(userId:string)
+  async getUserById(userId:string, exclude?: string[])
   {
-    return this.UserRepo.findOne({userId},{photo:true});
+    return this.UserRepo.findOne({userId},{photo:true}, exclude);
   }
 
   //create in the abstract repo only call .save() so
@@ -71,19 +77,20 @@ export class UsersService
         );
         
         newPhotoPublicId = cloudinaryResult.public_id;
-        
-        const photoEntity = new Photo({
+
+        const photoEntity = user.photo ?? new Photo();
+        Object.assign(photoEntity, {
           url: cloudinaryResult.secure_url,
           public_id: cloudinaryResult.public_id,
           filename: cloudinaryResult.original_filename,
-          mimetype: cloudinaryResult.mimetype,
+          mimetype: file.mimetype,
         });
-        
+
         user.photo = await this.PhotoRepo.create(photoEntity);
         Object.assign(user, updateUserDto);
         const result = await this.UserRepo.create(user);
-        
-        if (oldPhotoPublicId) 
+
+        if (oldPhotoPublicId && oldPhotoPublicId !== newPhotoPublicId) 
         {
           await this.cloudinaryService.deleteFile(oldPhotoPublicId);
         }
